+++ stories/SQForm.stories.js
@@ -186,7 +186,7 @@ export const BasicForm = () => {
           label="Ten Thousand Options"
           size={6}
           onInputChange={action('Update local state')}
-          lockWidthToField
+          lockWidthToField={false}
         >
           {MOCK_AUTOCOMPLETE_OPTIONS}
         </SQFormAutocomplete>